[Platinum V] Title: 고층 빌딩, Time: 0 ms, Memory: 6044 KB -BaekjoonHub

+++ 백준/Platinum/1328. 고층 빌딩/고층 빌딩.cc
@@ -3,7 +3,7 @@
 #define MOD_NUM 1000000007;
 using namespace std;
 
-long long int DP[101][101][101] = { 0, };
+int DP[101][101][101] = { 0, };
 
 int main() {
     ios_base::sync_with_stdio(false); cin.tie(NULL); cout.tie(NULL);
@@ -16,7 +16,7 @@ int main() {
     for (int i = 2; i <= N; ++i) {
         for (int j = 1; j <= min(L,i); ++j) {
             for (int k = 1; k <= min(R,i); ++k) {
-                DP[i][j][k] = (DP[i - 1][j - 1][k] + DP[i - 1][j][k - 1] + DP[i - 1][j][k] * (i - 2)) % MOD_NUM;
+                DP[i][j][k] = (DP[i - 1][j - 1][k] + DP[i - 1][j][k - 1] + (long long)DP[i - 1][j][k] * (i - 2)) % MOD_NUM;
             }
         }
     }
